Normative: Extend definition of "function code"
In the March 2018 meeting of TC39, the committee elected to extend the
definition of "function code" to include the BindingIdentifier of the
productions where a BindingIdentifier is present. This decision was
motivated by a desire to match implementation reality.

This alteration does not require a change to Test262 because like the
majority of modern implementations, Test262 already enforces the new
behavior [1].

[1] https://github.com/tc39/test262/pull/1404